test: colocate bootstrap artifact schema coverage
