+++ internal/agentartifacts/artifacts_test.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/Azure/unbounded/pkg/agent/bootstrapartifacts"
 	"github.com/Azure/unbounded/pkg/agent/goalstates"
 )
 
@@ -274,75 +273,6 @@ func TestCrictlArchive(t *testing.T) {
 			require.Equal(t, testCase.want, got)
 		})
 	}
-}
-
-func TestContainerImageArchivePath(t *testing.T) {
-	t.Parallel()
-
-	got := bootstrapartifacts.ContainerImageArchivePath("amd64", "mcr.microsoft.com/oss/v2/kubernetes/pause:3.9")
-	require.Equal(t, "container-images/amd64/mcr.microsoft.com_oss_v2_kubernetes_pause_3.9-a68ffa05fa78.tar", got)
-}
-
-func TestNormalizeManifest(t *testing.T) {
-	t.Parallel()
-
-	got, err := bootstrapartifacts.NormalizeManifest(bootstrapartifacts.Manifest{
-		Versions: bootstrapartifacts.Versions{
-			Kubernetes: "1.34.2",
-			Containerd: "v2.1.8",
-			Runc:       "v1.5.0",
-			CNI:        "v1.5.1",
-			Crictl:     "v1.34.0",
-		},
-	})
-	require.NoError(t, err)
-	require.Equal(t, bootstrapartifacts.Manifest{
-		SchemaVersion: 1,
-		Versions: bootstrapartifacts.Versions{
-			Kubernetes: "v1.34.2",
-			Containerd: "2.1.8",
-			Runc:       "1.5.0",
-			CNI:        "1.5.1",
-			Crictl:     "1.34.0",
-		},
-		ContainerImages: []string{},
-	}, got)
-}
-
-func TestNormalizeManifestPreservesContainerImages(t *testing.T) {
-	t.Parallel()
-
-	got, err := bootstrapartifacts.NormalizeManifest(bootstrapartifacts.Manifest{
-		Versions: bootstrapartifacts.Versions{
-			Kubernetes: "v1.34.2",
-			Containerd: "2.1.8",
-			Runc:       "1.5.0",
-			CNI:        "1.5.1",
-			Crictl:     "1.34.0",
-		},
-		ContainerImages: []string{" registry.example.com/pause:3.9 ", "", "registry.example.com/kube-proxy:v1.34.2", "registry.example.com/pause:3.9"},
-	})
-	require.NoError(t, err)
-	require.Equal(t, []string{"registry.example.com/kube-proxy:v1.34.2", "registry.example.com/pause:3.9"}, got.ContainerImages)
-}
-
-func TestNormalizeManifestRejectsUnsupportedSchema(t *testing.T) {
-	t.Parallel()
-
-	_, err := bootstrapartifacts.NormalizeManifest(bootstrapartifacts.Manifest{SchemaVersion: 2})
-	require.ErrorContains(t, err, "unsupported manifest schemaVersion 2")
-}
-
-func TestNormalizeManifestRequiresVersions(t *testing.T) {
-	t.Parallel()
-
-	_, err := bootstrapartifacts.NormalizeManifest(bootstrapartifacts.Manifest{})
-	require.ErrorContains(t, err, "manifest is missing required fields")
-	require.ErrorContains(t, err, "versions.kubernetes")
-	require.ErrorContains(t, err, "versions.containerd")
-	require.ErrorContains(t, err, "versions.runc")
-	require.ErrorContains(t, err, "versions.cni")
-	require.ErrorContains(t, err, "versions.crictl")
 }
 
 func TestCrictlVersionForKubernetesVersion(t *testing.T) {
